fetch-yarn-deps: add yarn-berry support
A new argument has been added to fetch-yarn-deps:
`yarnVersion` which will default to `1` which uses
the original `yarn`.

For newer `yarn-berry` `yarn.lock`, one can set the
`yarnVersion` to either `3` or `4` which will use
yarn-berry3 or yarn-berry4 respectively.

The difference is the `yarn.lock` file, which follows
a different format, depending on the version.

This also adds the corresponding tests.

The added support for yarn.lock files > version 1 has been
inspired by @szlend from
https://github.com/NixOS/nixpkgs/issues/254369#issuecomment-2080460150

fixes #254369

Signed-off-by: Florian Brandes <florian.brandes@posteo.de>
Co-authored-by: Doron Behar <doron.behar@gmail.com>